Centralize exact hard-equality checks in the constraints module.
Moves solution_satisfies_exact_hard_equalities out of the tier-step
executor so ladder finalize and tier logic share one canonical helper.
