+++ packages/api/api/analytics/military_score_inference/constraints.py
@@ -5,10 +5,12 @@ from dataclasses import dataclass
 from ortools.sat.python import cp_model
 
 from api.analytics.military_score_inference.accelerated_start import scoreboard_host_turn
+from api.analytics.military_score_inference.actions import ActionCatalog
 from api.analytics.military_score_inference.models import (
     CandidateAction,
     InferenceObservation,
     InferenceProblem,
+    InferenceSolution,
     ShipBuildCombo,
 )
 
@@ -192,6 +194,38 @@ class InferenceHardConstraints:
         aggregate_action_ids = frozenset(action.id for action in problem.aggregate_actions)
         if _fighter_transfer_actions_both_present(aggregate_action_ids):
             _add_fighter_transfer_direction_exclusivity(model, action_count_vars)
+
+
+def solution_satisfies_exact_hard_equalities(
+    solution: InferenceSolution,
+    observation: InferenceObservation,
+    catalog: ActionCatalog,
+) -> bool:
+    """Return whether a solution matches enforced hard equality targets."""
+    actions_by_id = {action.id: action for action in catalog.aggregate_actions}
+    combos_by_id = {combo.combo_id: combo for combo in catalog.ship_build_combos}
+    military_sum = 0
+    warship_sum = 0
+    freighter_sum = 0
+    for action in solution.actions:
+        catalog_action = actions_by_id.get(action.action_id)
+        if catalog_action is None:
+            return False
+        military_sum += catalog_action.score_delta_2x * action.count
+        warship_sum += catalog_action.warship_delta * action.count
+        freighter_sum += catalog_action.freighter_delta * action.count
+    for ship_build in solution.ship_builds:
+        combo = combos_by_id.get(ship_build.combo_id)
+        if combo is None:
+            return False
+        military_sum += combo.score_delta_2x * ship_build.count
+        warship_sum += combo.warship_delta * ship_build.count
+        freighter_sum += combo.freighter_delta * ship_build.count
+    return (
+        abs(military_sum - observation.military_delta_2x) <= observation.military_partition_slack_2x
+        and warship_sum == observation.warship_delta
+        and freighter_sum == observation.freighter_delta
+    )
 
 
 def observation_to_constraints_payload(
